Extracted code out to separate component
